progressing with dialogs
