+++ UI/Dialogs/DlgAppInfo.xaml
@@ -1,0 +1,48 @@
+﻿<Window x:Class="GolemUI.UI.Dialogs.DlgAppInfo"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d"
+        WindowStartupLocation="CenterOwner"
+        Title="App Information" Width="600" VerticalAlignment="Center" Height="300" ResizeMode="NoResize" SizeToContent="Height"
+        WindowStyle="None"
+        Background="Transparent"
+        AllowsTransparency="True"
+        x:Name="DialogAppInfo"
+     >
+
+    <Window.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="/UI/Themes/Defaults.xaml"/>
+                <ResourceDictionary Source="/UI/Themes/Borders/PageBackground.xaml"/>
+                <ResourceDictionary Source="/UI/Themes/WindowControls.xaml"/>
+                <ResourceDictionary Source="/UI/Themes/Buttons/GradientButton.xaml"/>
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>        
+    </Window.Resources>
+    <Border  BorderThickness="1" CornerRadius="12" Background="{DynamicResource SetupWindow.Background}" >
+        <Grid>
+        <StackPanel Orientation="Vertical" Margin="0">
+            <DockPanel>
+                <DockPanel x:Name="TitleBar" DockPanel.Dock="Top">
+                    <Button x:Name="CloseButton" Content="Close" DockPanel.Dock="Right" Click="CloseButton_Click" Style="{DynamicResource CloseWindowButton}"/>
+                    
+                    <TextBlock HorizontalAlignment="Center"></TextBlock>
+                </DockPanel>
+            </DockPanel>
+            <StackPanel Margin="30">
+                <Label Content="Raz" />
+                <Label Content="Raz" />
+                <Label Content="Raz" />
+                <Label Content="Raz" />
+                
+            </StackPanel>
+            
+
+        </StackPanel>
+            <Button Margin="0,0,0,20" HorizontalAlignment="Center" VerticalAlignment="Bottom" Width="300"  Height="45" Grid.Row="0" Grid.Column="4"  Name="btnRunBenchmark" Style="{StaticResource GradientButtonStyle}" Content="Close" />
+        </Grid>
+    </Border>
+</Window>
