--- UI/init.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:be2c1976da3e30c3655561acf604f57dc1f3258bab5a7984b8a0074934828a16
-size 495984
